pyproject.toml
[project]
name = "script"
version = "0.1.0"
description = ""
authors = [
    {name = "rian"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-storage (>=3.0.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
